html/semantics/embedded-content/the-video-element/video-loading-lazy-poster-when-visible.tentative.html

<!DOCTYPE html>
<link rel="author" title="Squarespace" href="https://www.squarespace.com/">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<style>
  #target {
    position: absolute;
    top: 1000vh;
  }
</style>
<body>

<script>
  const t = async_test(
    "Video with loading=lazy does not fetch poster while not visible, then fetches poster once scrolled into view"
  );

  const video = document.createElement("video");
  video.id = "target";
  video.loading = "lazy";
  video.poster = "/media/2048x1360-random.jpg?t=" + Date.now();
  const posterUrl = new URL(video.poster, location.href).href;

  document.body.appendChild(video);

  // Step 1: Not visible => check that poster hasn't loaded.
  t.step_timeout(() => {
    const entries = performance.getEntriesByName(posterUrl);
    assert_equals(entries.length, 0, "Poster should not be fetched when not visible");

    // Step 2: Scroll into view => poster should load.
    video.scrollIntoView();

    video.addEventListener("loadedmetadata", () => {
      const entries = performance.getEntriesByName(posterUrl);
      assert_greater_than(entries.length, 0, "Poster should be fetched after scrolling into view");
      t.done();
    });
  }, 2000);
</script>
</body>
