third_party/blink/web_tests/external/wpt/html/semantics/embedded-content/the-video-element/video-loading-lazy-poster-when-visible.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] Video with loading=lazy does not fetch poster while not visible, then fetches poster once scrolled into view
  assert_equals: Poster should not be fetched when not visible expected 0 but got 1
Harness: the test ran to completion.
